aider: Add index signature to TokenBoundAccount class.
# Aider chat conversation:

ASSISTANT: This error is occurring because TypeScript is not able to infer the type of the key when you're trying to access the properties of the `TokenBoundAccount` instance using bracket notation.

To fix this, we need to add an index signature to the `TokenBoundAccount` class. This will tell TypeScript that an instance of `TokenBoundAccount` can be indexed with a string, and the corresponding value will also be a string.

Here's how to modify your `TokenBoundAccount` class:

```typescript
src/resources/tokenBoundAccount.ts
<<<<<<< SEARCH
class TokenBoundAccount {
    client: Client;
    nftContract: string | null;
    destinationWallet: string | null;
    nftId: string | null;
    gasLimit: string | null;
    maxFeePerGas: string | null;
    chainId: string | null;
    privateKey: string | null;
=======
class TokenBoundAccount {
    client: Client;
    nftContract: string | null;
    destinationWallet: string | null;
    nftId: string | null;
    gasLimit: string | null;
    maxFeePerGas: string | null;
    chainId: string | null;
    privateKey: string | null;
    [key: string]: string | null | Client;
>>>>>>> REPLACE
```

This change adds an index signature `[key: string]: string | null | Client;` to the `TokenBoundAccount` class. This tells TypeScript that a `TokenBoundAccount` object can be indexed by a string, and the value at that index will be either a string, null, or a `Client` object.