aider: Add index signature to the `EstimateGasData` interface in `tokenBoundAccount.ts`.
# Aider chat conversation:

USER: Element implicitly has an 'any' type because expression of type 'string' can't be used to index type 'EstimateGasData'.
  No index signature with a parameter of type 'string' was found on type 'EstimateGasData'.ts(7053)

ASSISTANT: This error is occurring because TypeScript is not able to infer the type of the key when accessing properties of the `EstimateGasData` instance using bracket notation in the `estimateGas` method of the `TokenBoundAccount` class.

To fix this, you can add an index signature to the `EstimateGasData` interface. Here's how to do it:

```typescript
src/resources/tokenBoundAccount.ts
<<<<<<< SEARCH
interface EstimateGasData {
    privateKey: string;
    implementation: string;
    chainId: string;
    tokenContract: string;
    tokenId: string;
    gasLimit: string;
    maxFeePerGas: string;
}
=======
interface EstimateGasData {
    privateKey: string;
    implementation: string;
    chainId: string;
    tokenContract: string;
    tokenId: string;
    gasLimit: string;
    maxFeePerGas: string;
    [key: string]: string;
}
>>>>>>> REPLACE
```

This change adds `[key: string]: string;` to the `EstimateGasData` interface, which tells TypeScript that an `EstimateGasData` object can have any number of additional properties, and their keys are strings and their values are also strings.