aider: Add index signature to InitData interface.
# Aider chat conversation:

USER: now I'm getting this error: Element implicitly has an 'any' type because expression of type 'string' can't be used to index type 'InitData'.
  No index signature with a parameter of type 'string' was found on type 'InitData'.ts(7053)

ASSISTANT: This error occurs because TypeScript is not able to infer the type of the key when accessing properties of the `InitData` instance using bracket notation in the `init` method of the `TokenBoundAccount` class.

To fix this, we can add an index signature to the `InitData` interface. An index signature in TypeScript is a way to define the types you can use to index into the type, along with the corresponding return types when indexing.

Here is the change you need to make:

```typescript
src/resources/tokenBoundAccount.ts
<<<<<<< SEARCH
interface InitData {
    privateKey: string;
    destinationWallet: string;
    chainId: string;
    gasLimit: string;
    nftContract: string;
    nftId: string;
}
=======
interface InitData {
    privateKey: string;
    destinationWallet: string;
    chainId: string;
    gasLimit: string;
    nftContract: string;
    nftId: string;
    [key: string]: string;
}
>>>>>>> REPLACE
```

This change adds an index signature to the `InitData` interface, which allows TypeScript to infer the type of the key when accessing properties using bracket notation.