aider: Modified the index signature in `src/resources/tokenBoundAccount.ts` to allow functions that take an argument and return a Promise.
# Aider chat conversation:

USER: can you help me correct my `src/resources/tokenBoundAccount.ts` that has a warning on line 69 that says: Property 'init' of type '(data: InitData) => Promise<void>' is not assignable to 'string' index type 'string | Client | (() => Promise<string>) | (() => Promise<any>) | (() => Promise<void>) | (() => Promise<void>) | null'.ts(2411)

ASSISTANT: Sure, I can help with that. The issue is that the index signature in your class is too restrictive. It's currently set to allow properties that are either a string, Client, or a function that returns a Promise. However, your `init` method is a function that takes an argument and returns a Promise. This doesn't match the index signature.

To fix this, we need to modify the index signature to allow functions that take an argument and return a Promise. Here's how you can do it:

```typescript
src/resources/tokenBoundAccount.ts
<<<<<<< SEARCH
[key: string]: string | null | Client | (() => Promise<string>) | (() => Promise<any>) | (() => Promise<void>) | (() => Promise<void>);
=======
[key: string]: string | null | Client | ((data?: any) => Promise<string>) | ((data?: any) => Promise<any>) | ((data?: any) => Promise<void>) | ((data?: any) => Promise<void>);
>>>>>>> REPLACE
```

This change will allow the `init` method to be assigned to the class without TypeScript complaining. However, please note that this change will also allow any function that takes an optional argument and returns a Promise to be assigned to the class. If you want to restrict this further, you would need to define more specific types for the functions.