aider: Update types in `types/types.d.ts` to reflect properties and methods in `Dippi` and `TBA` classes in `src/dippi.ts`.
# Aider chat conversation:

ASSISTANT: Yes, I see a need to update the types in `types/types.d.ts`.

The `Dippi` and `TBA` classes in `src/dippi.ts` have properties and methods that are not reflected in the `Dippi` and `TBA` interfaces in `types/types.d.ts`.

For example, the `Dippi` class in `src/dippi.ts` has properties `auth`, `user`, `wallet`, `application`, and `applicationToken` which are not present in the `Dippi` interface in `types/types.d.ts`.

Similarly, the `TBA` class in `src/dippi.ts` has properties `auth` and `_tokenBoundAccount` and methods `init`, `create`, `estimateGas`, and `createAccount` which are not present in the `TBA` interface in `types/types.d.ts`.

Here are the changes that need to be made:

```typescript
types/types.d.ts
<<<<<<< SEARCH
    class Dippi {
        constructor(config: DippiOptions);
        setAuthToken(newAuthToken: string): void;
        auth: any;
        user: any;
        wallet: any;
        application: any;
        applicationToken: any;
    }
=======
    class Dippi {
        constructor(config: DippiOptions);
        setAuthToken(newAuthToken: string): void;
        auth: AuthInterface;
        user: UserInterface;
        wallet: WalletInterface;
        application: ApplicationInterface;
        applicationToken: ApplicationTokenInterface;
    }
>>>>>>> REPLACE
```

```typescript
types/types.d.ts
<<<<<<< SEARCH
    class TBA {
        init(args: InitArgs): Promise<any>;
        create(): Promise<any>;
        estimateGas(): Promise<string>;
        signCreateTransaction(): Promise<string>;
        sendCreateTransaction(): Promise<string>;
        setAuthToken(newAuthToken: string): Promise<void>;
    }
=======
    class TBA {
        auth: Auth;
        _tokenBoundAccount: TokenBoundAccountInterface;
        init(args: InitArgs): Promise<any>;
        create(): Promise<any>;
        estimateGas(): Promise<string>;
        createAccount(): Promise<any>;
        setAuthToken(newAuthToken: string): Promise<void>;
    }
>>>>>>> REPLACE
```

Please note that the types `AuthInterface`, `UserInterface`, `WalletInterface`, `ApplicationInterface`, `ApplicationTokenInterface`, `Auth`, and `TokenBoundAccountInterface` should be defined somewhere in your project. If they are not, you will need to define them.